library: match dynamic imports with exported symbols
This first prototype takes a full path to a library and tries to
resolve all symbols which are currently undefined to the matching
functions in the imported libraries. It uses a worklist to resolve
unknown symbols recursively, and caches the results based on the
full path to the respective library.

There are many things left to do:
- Symbol versions are currently ignored
- The distinction between STT_FUNC and STT_IFUNC (which is defined to
    the same value as STT_LOOS as it is OS specific) is not
    implemented, which makes it miss some functions from glibc.
- The resolution of the exports is not written back to the library
    objects yet, the loop at the bottom only prints them.
- We could do statistics about things like percentage of used exports
    and similar metrics
- Provide a generic interface to scan the whole system