store: change symlink allocation
So far, a symlink was only added into the store if the underlying
actual library was already present. Otherwise, a Library object was
created using the symlink's path (which we then tried to fix up in
_resolve_libs). This caused errors for some libraries with DT_RPATHs
which were created from their symlink, as library.parse_functions
used the symlink's path to replace $ORIGIN instead of the actual path,
which lead to imports not being found.

This change changes the allocation: _get_or_create_library always
uses a realpath to build the Library object and offers an additional
return value with the name of the corresponding symlink. This link can
be added to the store later, as we might need to do additional checks
(e.g., target.is_compatible(needed) in _find_compatible_libs).