directory: resolve functions in a flat directory
The libraryopt tool allowed users to provide a directory which
contains all relevant libraries which was then scanned and cross-
resolved functions inside the libraries.

Our LibraryStore module, on the other hand, only works on a running
system as it relies on the output of ldconfig to locate the .so files
for a given NEEDED tag in an importing library.

This new script cross-references imports and exports for libraries in
a given directory and allows the resolution to work without consulting
ldconfig.

There are still some warnings about _init/_fini being defined more
than once and there is no real conflict resolution yet for similar
cases, but it's a start.