ldresolve: fallback search for libraries by name
On our Debian stretch installation, the /usr/bin/ldap* binaries
have a dependency on libldap-2.4.so.2, but the output of
'ldconfig -p' only knows libldap_r-2.4.so.2 (notice the missing
'_r').

At this point, the system search paths of ld.so come into play, as
can be seen from the output of the following command:
'LD_DEBUG=libs /lib64/ld-linux-x86-64.so.2 --list /usr/bin/ldap*'

As a quick fix, implement a fallback mechanism in LDResolve which
looks for the required file name in all directories which appear
in the output of 'ldconfig -p'. This is not the full system search
path but this one is much harder to cleanly extract.

The fallback mechanism is only invoked if the regular resolution
using the results of LDResolve.get_paths is not successful.