store/library: gather all recursively loaded libraries
In Library.needed_libs, only the libraries marked as DT_NEEDED by
the ELF file itself are gathered. In some libraries, functions are
used which are only present in recursively loaded libraries (that
is, libraries which are DT_NEEDED in a library loaded by the ELF
file in question).

Create a new attribute in Library which is updated with all
DT_NEEDED libraries from all child libraries once they have been
processed. This attribute is also exported to JSON and now used in
resolve_all_functions to locate more functions correctly.