store: also resolve symbolic links in dirname component
The check for os.path.islink() only returns True if the basename
(that is, the filename of the library to analyze) is a link itself
but not if the directory leading up to it contains a linked
component. This happens for example with /usr/bin/X11 which is
often linked to /usr/bin.

In this case, the resolution of imports could not localize imports
using RPATHs relative to the original binary as it did not use the
actual location.