store: parallel version of library resolution
The recursive evaluation process of individual libraries
is an optimal target for parallelization - we can resolve
the needed libraries independently on many cores, and still
use a shared dictionary as a caching mechanism, so that the
evaluation on other cores can profit from already processed
libraries.

In order to parallelize the evaluation, this change breaks
the evaluation out of the LibraryStore class and makes the
required methods global functions instead. The functions
now take an additional argument (store) instead of 'self',
so the new process_from_list can pass a shared dictionary
to the functions.

Additionally, '_find_compatible_libs' does not fill the
'all_imported_libs' dictionary anymore as the order can
change depending on the evaluation order on other cores.
Instead, the new 'fill_all_needed' function reconstructs
'library.all_imported_libs' after the recursive evaluation
has finished (this takes under a second on my test
machine).