scripts/running_analysis: parameter for function resolution, store results
In order to crossreference all imported functions to their respective
paths, add a new parameter which triggers the resolution. Additionally,
the results are now added into the Library objects (as values in the
respective imports dictionary) and subsequently stored in the output
file. To see when the referencing starts and ends when running in
verbose mode, two logging.info() statements have been added.