store/scripts: Parameter for executables as resolution entry points
On a running system, we can gather all libraries from the output
of running ldconfig -p and using the base paths of all libraries
as directories to search. When we're interested in the usage of
libraries, however, we might want to start at executables only:
using all libraries as starting points would mark functions as
used which are only referenced from another (possibly unused)
library but not even transitively from a binary.

This adds a new parameter '-a' to start from all known libraries,
the default is False, meaning we start at executable files only.