store: get rid of result dict in resolve{_all}_functions
After introducing propagate_call_usage, passing around the
result from resolving functions in resolve{_all}_functions
became a bit unhandy as we need to keep a reference around in
the caller (i.e., running_analysis.py).

The same information can always be recreated from the objects
in the LibraryStore, so we can get rid of storing it and
passing it around, and (re-)build the mapping when we need it.