library: add support for calls through .plt
Shared libraries might not only call functions by their offset
inside the shared object, but also through the Procedure
Linkage Table (PLT). Mostly, this mechanism is used for calls
into other shared libraries in order to invoke the loader to
resolve the actual relocated address of the symbol in
question.

This mechanism can also be used for calls inside the same
library. In that case, our normal mechanism to identify calls
to other interface functions fails as the jump/call target is
the address of the .plt entry, not the actual function body.

This change adds support for calls through the .plt including
test cases for capstone and objdump disassembly.