library: don't fail on corrupt .dynsym entries
When analysing machines in our central computer rooms, I came
across a library (libatlas.so.3.0) which exports and uses
functions which are listed with STT_NOTYPE instead of STT_FUNC
in the .dynsym section.

This makes our call propagation crash as the exports dict
does not contain those functions after parsing. For now, log
an error message and ignore these functions.