store, running_analysis: provide additional entry points via parameter
With the introduction of the '-a' parameter, we already had
the choice of starting the resolution of functions and the
propagation of function users from either all ELF objects or
only the ones which are executable.

As we do not have a way of determining dynamically loaded
libraries yet (i.e., libraries loaded using dlopen()), we may
need to provide those libraries as entry points for the
search to at least overapproximate their impact on library and
function usage further down the dependency chain.

As an example, Apache and PHP ship dynamic libraries which
provide additional functionality - these are loaded on demand
with dlopen(). As we can not statically determine which ones
are actually used, we could include all of them (or a known
set of used ones) into the list of additional entry points
in order to have them listed as users of their imported
libraries.