store: add recursive calls to transitive set
If we had a function calling itself (either directly or
through another function) it didn't show up in the result of
store.get_transitive_calls() as we wanted to avoid multiple
descends into the recursive call. This, however, also
excluded the recursive function from showing in the result.

Fix it by adding it to the result before checking if we are
already working on it.