library: enable generation of uprobe events
The main script now has a parameter to automatically generate
uprobe events[0] for all libraries in the LibraryStore.

The generation of event strings uses offsets inside files. To
allow dynamic tracing of functions inside executables, we
need to subtract the load address from the value of every
symbol as these reference addresses in the loaded form - for
dynamic libraries and PIE executables the offset should be 0.

Additionally, if a .symtab section is present, we extract all
offsets for local functions as well to get a more fine-grained
picture during the dynamic analysis. For executables we also
add the entry point address from the ELF header to the set of
probe-able points.

[0]: https://www.kernel.org/doc/Documentation/trace/uprobetracer.txt