store: use all_entries parameter for uprobe generation
As the other methods of LibraryStore dealing with Library
objects have an 'all_entries' parameter to distinguish between
working on all ELF files and only executable ones, let's use
it here as well.

However, default to 'True' as the static analysis might have
missed something we want to discover during the dynamic
analysis.