scripts: add tracing for dlopen/dlsym
This adds a systemd service and corresponding helper script
for tracing information about dlopen-ed libraries and
functions which are read from them via dlsym.

The parse_trace.py script reads a trace file and correlates
the dlsym calls to the libary in question. It generates a
list of <library>:<function> pairs as its output which - as
a file - can be passed to the running_analysis.py using the
'-u'/'--used-functions' parameter. These functions will then
be marked as used by the analysis and propagate their usage
information into referenced library functions.