running_analysis: check existence of additional entry points
The paths provided by the '-e' and '-u' parameters were not
checked so far, which could lead to errors further down the
analysis (for example, as LibraryStore.get_from_path would
return 'None' for an entry point).

Fix this by using os.path.isfile() for every path in question
and checking if the added paths actually resolve to a library
in the relevant functions in LibraryStore.