kernel-patches: update for upstream fix
Ubuntu picked an upstream update to trace_uprobe.c for a
newer release of the 4.15.0 series as commit 733687f82671
("tracing: Fix bad use of igrab in trace_uprobe.c"). This
fix changed code close to our changes and hence broke the
application of the second patch.

This change updates the patch to a rebased version on top
of 4.15.0-34.37.