library: prevent crash on invalid section headers
If the ELF headers of an analyzed binary are corrupted,
they might be interpreted as if no sections are present in
the file. This lead to an exception in Library.__init__ as
we didn't check the result of get_section_by_name('.text').

Fix this by raising an ELFError if no text section is
found.