library: support versioned symbols
In order to keep stable interfaces for programs, libraries
can use symbol versioning to provide multiple different
implementations of the same function or add new functions
which are then only provided in a newer version of the
library.

The version information is not part of the "normal" name
of a symbol but connected to it through the
.gnu.version{,_d,_r} sections. readelf displays them next
to the symbol name in the .dynsym/.symtab section in the
form <name>@@<version> (e.g., printf@@GLIBC_2.2.5).

This change makes our tool aware of versioned symbols.