store: fill all_needed_libs in breadth-first order
So far, the all_needed_libs dictionary was updated directly
after descending into the needed library, leading to a
depth-first order of libraries in all_needed_libs.

The ELF specification however states that symbols are
resolved in breadth-first order. This changes the results
for ELF files which use symbols which are defined in
more than one of the (transitively) imported libraries.