store: improve readability of dump/load
We had a lot of identical code for dumping and loading
Library objects which was hard to read and got worse with
every additional member of the Library class.

This change introduces local helper functions to help with
the conversion of OrderedDicts and dicts containing sets as
their values into/from their JSON-compatible
representation.