store: fix Python 2 compatibility
The 'symlink'-like entries in LibraryStore are of type
'unicode' in Python 2 and of type 'str' in Python 3.

This made no difference before pyelftools was updated to
properly handle Unicode strings. With our update of
pyelftools in commit ec7b7b8a053e we imported these
changes and hence broke compatibility with Python 2.

As there is no clean way to unify these types and properly
check for them in a way compatible to both Python versions,
let's turn the condition around and check for Library
objects instead.