library: speed up and simplify symbol version lookup
elffile.get_section_by_name() always creates a new Section
object but we only need that to be constructed once. Let's
add a mapping for resolving symbol versions:

Library._version_names maps an index (which is the symbol
index in the associated .dynsym table) to the name of the
associated version (either defined or required).