library: fix parse_plt for 32 bit libraries
Library.parse_plt() only looked for a '.rela.plt' section
to correlate the relocation entries with the PLT, but in
32 bit binaries the relevant section is called '.rel.plt'.

Additionally, the sh_entsize member of the .plt section
header is often set to 4 in 32 bit systems even though the
individual PLT entries are still 16 bytes long. According
to https://reviews.llvm.org/D9560 a valid solution for this
mismatch is to round the sh_entsize up to the value given
in sh_addralign.