store/library: implement loader-like behaviour for weak functions
In this change a new mode is added for function resolution,
triggered by the '--loaderlike' parameter for the analysis of
a full system.

The loader will always look up functions starting from the
main binary. This means, a library importing 'malloc' may not
necessarily use the implementation from one of its own
imported libraries - if the binary provides an implementation,
the loader will instead use that definition for all further
imports of the function.

In combination with weak symbols, we also need to keep track
of the type of every symbol. Weak functions can be overridden
by strong symbols but if only weak definitions are found the
first one will be used.

This change also improves LibraryStore.propagate_call_usage()
by introducing a combined worklist containing tuples of the
form (Library, function address). This means we can
selectively add functions to the worklist instead of
re-evaluating a whole library if the users of one function
have been updated.

Furthermore, we can now also use external ELF files for local
function detection if build-id is not used but the file is
placed under '/usr/lib/debug/<path>' instead. build-id is
still preferred if present.