librarystore: loaderlike with all dependencies added
The loaderlike mode relied completely on detected calls in
imported functions to propagate usage data into the imported
libraries (except for the binary where all detected global
functions and all functions directly imported by it are
marked as used).

If we miss edges in the call graph this mode might become
quite inaccurate. If we're only interested in how weak
functions affect the analysis we might also want to use
the loaderlike mechanism in principle but also mark all
referenced functions as used (which is what the normal
transitive resolution does).

Allow this by introducing an additional parameter to the
loaderlike resolution to force addition of the importing
library into the imported function.