library: keep track of users of local functions
In order to remove local functions from libraries as well, we
should be keeping track of users of those functions which are
propagated into and through them instead of ignoring them.

With this we can also generate separate statistics for
exported and local functions and for the combination of both.