library: support searching for .rela.plt and .plt in ultra-stripped ELF files
In ultra-stripped binaries (that is, ELF files without a
valid section table) we do not know where the .rela.plt and
.plt sections are in order to find calls to imported or
exported functions.

However, the .dynamic segment has to be present and it
contains information for the .rela.plt section (the number
of relocations as well as the size of the individual
entries and the beginning of the section). For the .plt
section, we have to traverse the binary to find the 'magic'
first entry which always has to follow a certain pattern.

Note that this might be unreliable and we only use it if
the normal method of finding the sections fails.