library: take maximum size for conflicting ranges
When parsing the .symtab of a binary/library we can have
multiple entries for the same address but some entries could
have a size of 0 which would override a bigger value read
from a valid entry earlier.

Let's always take the maximum of all read values for the size
of a function.