interface_calls: allow to pass Library objects into multiprocessing
For large binaries, constructing the Library (and thus,
ELFFile) object can take quite a lot of time, so we might
want to do this only once. So far we only passed all paths
into the multiprocessing Pool when searching for calls so
that every parallel process had to rebuild the Library
object before disassembling.

However, all members of the Library object (except the
ELFFile) are serializable, and the underlying ELFFile can
easily be closed after our processing step while filling the
Library object. This means we can simply pass the Library
objects into the Pool and get rid of the overhead of
re-evaluating the whole ELFFile before disassembling it.