interface_calls: heuristic for function pointer access
On x86_64, a pointer to a function is mostly taken by
calculating an offset from the current instruction pointer
and loading that address into a register via the 'lea'
instruction. When we disassemble a function, we can
calculate the value of that pointer and check if the
corresponding address is the address of a local or global
function.

This also means that the function we're currently looking
at has a dependency on the target function, so we should
keep track of that dependency for the transitive propagation
of usage information in a later step. Let's add the target
address to the corresponding dictionary in the Library
object.