library/interface_calls: detect more types of calls
Some handwritten assembly code might jump to a label
somewhere in the middle of an actual function. This
situation was missed so far as we only checked for jump/call
targets against all known starts of functions. With this
change, we can also check if a jump or call lands inside
a known range.

Additionally, some binaries linked with LD_BIND_NOW might
directly use the .got entry to call/jump to an imported
function. In order to detect these, we need to parse the
.rel{,a}.dyn relocations and check for references to
imported functions from there. In the call resolution, we
can treat them just like calls through the .plt.