library, store: mark _init and _fini as used
The special functions _init and _fini are called upon library
initialization and destruction. As function calls can happen
from them, too, let's mark them as required during call
propagation.

Additionally, the entries in .dynsym or .symtab might only
have a size of 1 while the functions are in fact bigger. To
fix this, take the maximum of the surrounding section (.init
or .fini, respectively) and the size reported in the symbol
table.