store: fall back to brute force search if functions are not found
The function resolution might fail for some symbols if we are
missing some dynamic dependencies or a DT_RPATH setting of a
top-level binary. In order to still resolve functions to their
origin, add a brute-force mechanism which searches for a
definition of an unknown symbol in all known libraries. This
step is only taken if both versioned and unversioned lookups
have failed.