library, store: enable tracking of more function pointers
So far, we could only track references to function pointers
if they were used as callbacks inside the code on x86_64,
that is, if a RIP-relative access to an exported or local
symbol was detected.

However, function pointers are often used in global variables
or structs to build up objects, which in turn represent an
abstraction of the implementation below. Furthermore, C++
code using virtual methods uses so-called vtables, which
also hold pointers to methods and are switched dynamically
if the dynamic type of the underlying object changes (i.e.,
an object is assigned or cast to a pointer of a super class).

Such tables need to contain actual addresses of functions as
their members will be loaded into a register and called from
there. In shared objects, however, the address is only known
at load time, when the dynamic linker/loader has decided
where to put the shared library. To fix these places up
automatically, ELF uses relocations which tell the loader
'please put the actual address of symbol X to address Y'.
This address Y is mostly placed in the global offset table
(GOT), and code references these GOT entries to access the
underlying structs and functions.

Through additional processing of the relocation sections, we
can also reconstruct the placement and dependencies between
objects and functions. This change adds the necessary logic
to the Library class, collects references to objects while
disassembling each function, and gathers and propagates
transitive dependencies through object references in the
LibraryStore.

Now, we mark all functions which are referenced from a certain
object as used if a function interacts with the object. This
also works transitively, meaning that nested pointers to
structs are also evaluated and followed.

While the changes for x86_64 are rather straight forward as
RIP-relative addressing is also used to access objects through
the GOT, i386 is a bit tougher: here, the compiler needs to
dynamically load the current instruction pointer into a
register (using a function like __x86.get_pc_thunk.ax when
using gcc) and calculate on that register. We also added
a heuristic to determine the value of the respective register
and update the value until it is overwritten. Doing so
allows us to capture accesses to GOT entries through the thunk
register.