library: use bisect to determine if relocations point into objects
So far, we naively iterated over all known objects while
processing the relocation section in order to determine if a
relocation points inside another object. For ELF files with
many objects, this slows down the parsing process. As a
mitigation, we can sort the list of objects by their start
address once (before processing the relocations) and look up
possible intersections by using bisect.bisect() with the
target address and checking if the previous object in the
sorted list encloses this target address.

This significantly speeds up processing for large binaries
like libinkscape_base.so or the dockerd binary.