store: use global caches for objects and transitive function calls
LibraryStore.get_transitive_calls() used a local cache for
every direct call, and only initialized and filled it for
subsequent recursive calls. For huge binaries, this slows
down the analysis significantly, so let's keep the cache
alive across invocations. Additionally, use a cache for the
functions which are transitively referenced through objects
to speed up this part of the propagation as well.