library: add PLT links for IRELATIVE relocations
So far, relocations of type R_X86_64_IRELATIVE were skipped
by the code in Library.parse_plt(). However, this makes us
lose connections from calls to these IRELATIVE functions
inside the shared library.

Add links from the PLT entries to their corresponding
implementations as well. For now, this is only implemented
for x86_64 but will be adapted for i386 in the future.