store: load default values if member not in stored JSON file
Up to now, if we added a new member to the JSON store and
added the code to load this member from a saved version,
loading older versions of a store would fail as the JSON
file would not include the new members and access to them
would lead to a KeyError.

Let's make this more robust by initializing the members of
the Library objects with default (empty) lists or
dictionaries, depending on the loaded member.