librarystore: implement STB_WEAK lookup logic in _find_imported_function
So far, our function lookup stopped when it found the first
matching definition for a function. In reality, however, the
loader keeps going until it either finds a definition with
visibility STB_GLOBAL or runs out of libraries to look in.

Implement this as well, and match the glibc loader's
behaviour by keeping the first STB_WEAK definition if no
STB_GLOBAL definition is found.