library, store: rework function resolution to match glibc loader
Resolution of symbols in the glibc loader does not keep a
list of already defined symbols to look things up but rather
starts the process of looking symbols up in a linked list of
maps which describe the involved ELF files in the order they
are loaded. For every already loaded file, the same rules
regarding symbol versioning apply for weak and overridden
symbols which we could not properly emulate using our
previous approach.

Rewrite the lookup process to resemble the actual loader
behaviour more closely. Note that this applies to objects in
the same manner, this will be implemented later.