library: store maximum size of objects at the same address
In some shared libraries we found objects which start at the
same address but have a different size, depending on the
version of the associated symbol. For example, if later
versions of a global struct added members, these are only
accounted for in the st_size value for the later symbol
versions. So far, we always used the size of the version we
encountered last but this could lead to missing function
pointers added only in later versions.

In order to include those, rather keep the maximum size of
all objects at the same address. Note that this does not
necessarily match the exact requirements of an executable
which imports an old version (as this old version should
optimally not include pointers added later on), but our
resolution process can not reflect this (yet). With this
solution, we play things safe and possibly keep a few
unneeded functions but do not remove possibly required ones.