library: support external debug directories via environment
So far, external .symtab sections could only be looked up in
the place where a currently running system would install
them, namely in the /usr/lib/debug/{,.build_id} directories.

As we might have to analyze libraries which were copied from
somewhere else, add two environment variables which specify
a place where debugging files are stored. You can provide
EXTERNAL_DEBUG_DIR as the base directory where files are
stored using the regular name (for example, if we search for
debugging symbols for '/lib/libc.so' and EXTERNAL_DEBUG_DIR
is set to '/tmp/bar', we will check if the debugging file is
at '/tmp/bar/lib/libc.so'.

Additionally, if the debugging information is provided via
the build ID, you can provide EXTERNAL_BUILDID_DIR as the
base directory in which the folders and files generated from
the build ID hash are located.