library: allow marking extra local functions by name
In order to allow marking of traced functions across versions
of libraries, we should not rely on the offset of the function
in the binary as this will most definitely change when the
library is recompiled.

Let's also use their names instead, and mark all instances of
local functions with the same name as used.