uprobes: fix for name size limitation on newer kernels
Newer Linux kernels had their uprobe tracing code merged with
the more general dyn_event infrastructure which also means
that a uprobe event can no longer have an arbitrarily long
name but is limited to MAX_EVENT_NAME_LEN characters which
currently is set to 64. With our current implementation,
longer names could easily be generated from using the full
path to the library in the event name.

Instead, use a counter for the number of events, generate an
event named 'trace_probe_<counter>' and improve the uprobe
collection script to read the mapping from event number to
the actual path and offset.