library: rearrange _init and _fini handling
So far, our code only dealt with discrepancies between the
size of the _init function and the size of the .init section
while parsing .dynsym. However, _init and _fini can also be
symbols with bind STB_LOCAL, so we need to apply the same
logic in parse_symtab() as well. Additionally, we missed the
_fini function so far while parsing .symtab, so let's also
add this.