store: propagate users into local and exported objects
So far, we only added users to exported objects if the
references were between different shared objects. However,
an object or variable will also be referenced from functions
inside the respective binary, so we need to propagate users
into the object_users dictionary.

External object references will already be marked by the
resolution step but we might also implement this later in
the propagation step.