store: also resolve objects in resolve_functions_loaderlike
Our current implementation of the loaderlike behaviour
ignored objects when resolving dependencies between shared
libraries. This did not show up in our testcases as the
Ubuntu installation used in our lab was using -Wl,--as-needed
as a default linker flag while the newer Debian config
does not set it.

Fix this by including the "simple" straight lookup in the
list of libraries to be loaded. If required, this could also
be ported to use the actual versioning behaviour from
_find_imported_function later on.