interface_calls: distribute work at function granularity
Instead of giving a whole library to one worker thread in
the thread pool, generate individual functions as work
packets. This might incur a little more overhead for small
numbers of small libraries, but greatly reduces the time
required for 'asymmetric' cases where we would otherwise
be stuck in a de facto single-thread mode for one large
(and many small) libraries.