library: postprocess function ranges to include trailing NOPs
The size information from the .dynsym or .symtab tables do
not include any NOPs that are appended to the end of a
function (for example, to achieve correct alignment of the
following function). In our case, this leads to small bits
of NOPs between identified functions. In order to get better
coverage of continuous segments of code, let's add these
NOP bytes to the range of the preceding function.

As we might also have gaps in the binary from unidentified
functions (when we're missing .symtab, for example), we only
add the NOP bytes if the whole gap consists of NOPs.