library: add more possible locations for external debug files
When files with debug information are collected in a single
flat directory, we could not use them as we always appended
the full path to the library to the provided debug directory.

Let's also add the path consisting of the provided directory
and the basename of the library to the paths which are
checked for a corresponding .symtab.