library, interface_calls: detect overlapping of symbol ranges
In some glibc versions, local symbols were contained within
the range of exported symbols (e.g., __nanosleep_nocancel was
completely inside __nanosleep). Without proper detection, this
could lead to a crash when overwriting the inner function part
when the static analysis failed to detect the connection
between the two.

Fix it by connecting the outer and inner function in the call
graph. While at it, simplify the code for the disassembly
dictionary handling.