library: only try to fix NOP gaps if .symtab information is known
If we don't have any information about local functions we do
not even have to try to close gaps between functions. The
gaps will be very large and possibly take a long time to
disassemble only to discover they contain instructions other
than NOPs.

To fix it, make parse_symtab return a boolean if a symbol
table was found and only postprocess the ranges if that's the
case.