library, interface_calls: keep track of calls through RELATIVE relocations
RELATIVE relocations might point into globally or statically defined
structures which was already noted before. However, we did not check
for any direct calls through the address of the relocation when
reconstructing the call graph from disassembly. To achieve this,
store the RELATIVE relocations which point to local or exported
functions and check if any call or jmp operations use the address of
the relocation. In this case, add an edge to the original function.