store: mark ELF entry point as used when propagating
We already added the 'main' function and any initialization
functions to the propagation worklist, however, execution of
an ELF binary starts earlier - to be precise, at the address
stored in the 'e_entry' field of the ELF header. The function
at this address will mostly call __libc_start_main with the
main function as a parameter but it can also call other
imported functions.

Let's try to locate the corresponding local or global function
to the address, mark it as used and add it to the initial
worklist.