library: filter out init/fini functions with address 0
This most likely means that the linker did not put an offset
into the read location as offset 0 should contain the ELF
header of the file. This situation can arise from symbols in
the INIT_ARRAY which are resolved by the loader via a
R_X86_64_64 relocation, for example. I'm working on doing
another pass over the relocation table to properly find
the corresponding relocation entries and add them to the
Library.init_functions list.