library: check for relocations into INIT_ARRAY
Some entries in the initialization array pointed to by the
DT_INIT_ARRAY member in the dynamic section are only filled
by relocations (and do not have a relocatable offset saved
at the corresponding address). To cover this case, let's
add another iteration over the relocations in .rel(a).dyn
and check if any R_X86_(64_64, 32_32) relocations are
pointing into the initialization array. If we find such
relocations, add their target to the list of init functions.

As an example, this situation occurs with the
__cpu_indicator_init function in libgcc_s.so.1 from Debian
Buster.