running_analysis, library: allow hex and pattern matching for functions from allow list
So far, we could only exactly match a function name or take
its address as a decimal number. As looking up functions via
readelf or gdb will often provide us with a hex number, let's
also take them by checking for a '0x' prefix and converting
the number accordingly.

Additionally, it could be useful to not only match a single
function by name but provide a regular expression to be
matched in order to add groups of similarily named functions
all at once.