interface_calls: avoid overhead from redundant sorting
If the target of a call or jump instruction does not match a
known local or global function, we try to determine if the
target is inside the range of another function. This might
happen if hand-coded assembly jumps over some initialization
code, for example. However, the current implementation sorted
the list of function ranges every time it encountered this
scenario, even though the function ranges do not change at
this stage of processing. This causes massive overheads for
large binaries as the same list of thousands of elements is
sorted over and over again.

Instead, sort the list once before the parallel disassembly
step starts and use the pre-sorted list in the checks. For
large libraries like libLLVM-7.so.1, we noticed a 28x
speedup for the disassembly step.