library: remove logging.debug in hot path
As Library.add_export_user is called a lot from the usage
propagation step, we can see quite some overhead from the
logging call during profiling. As this is not strictly
necessary, let's remove the two calls. Additionally, the
usage dictionaries do not need to be ordered so let's
convert them to native dictionaries for a small performance
gain.