librarystore/uprobes: automatically strip mount prefix
When we're analysing a file system image from another host
(e.g., a mounted OpenWRT partition), we still want to
generate the corresponding uprobes to match the path inside
the target file system. So far, this required manual editing
of the generated uprobes file twice (once before copying the
file to the target system and reverting the changes when
mapping the uprobes back to functions in
parse_collected_uprobes.py).

Now, the MOUNT_PREFIX environment variable is automatically
stripped and re-applied at the respective points.